requirements.txt
torch torchvision torchaudio
pillow
pandas scikit-learn
fastapi uvicorn[standard]
onnx onnxruntime
numpy
